Streamline local statics
Summary:
I noticed that local statics in closures were very inefficient - it
did a function call, which dynamically constructed the name of the
property, dynamically looked it up, boxed the resulting property if
necessary, and then set the inner cell to uninit as a flag which the
caller had to test.

Instead, figure out the property offset at jit time, and do a load and
add to get the property address. We can then test for uninit, and box
it on the slow path.

The non-closure case was much more optimized (no calls on the fast
path), but still generated some pretty bad code. This attempts to
streamline it all.

Reviewed By: ricklavoie, hubyrod

Differential Revision: D5217786

fbshipit-source-id: 52b8b995f40f120aa110e36548fbea0c5085c007