More phi-optimizations
Summary:
If a DefLabel has a single output, and all of its inputs are
constants, and its output is fed directly into a Jmp[N]Zero and not
otherwise used, we can just redirect the incoming Jmps to the
appropriate successor blocks.

In the common cases where we produce such phis, the value was a
boolean stored on the stack, so there's actually an StStk between the
DefLabel and the Jmp[N]Zero, so we need to do this after
store-elim. But the main benefit of this optimization is that can
enable futher store-elim opportunities; so restructure the opts pass
to keep looping while optimizePhis manages to improve things.

Reviewed By: ricklavoie

Differential Revision: D5217882

fbshipit-source-id: 9f5e943cf49bc4e3119cc34777f472eaba8bf65c