Small tweak to refcount-opts
Summary:
First, two identical LdStaticLoc's always refer to the same Ref, so
put them in the same must-alias sets.

Also, if all the inputs to a DefLabel are in the same must-alias set,
the output should also be in the same set.

Finally we can move two increfs through a phi if there's a suitable
ssatmp in the phi's block. The previous restriction was that at least
one of the input ssatmp's was available in the phi's block.

Reviewed By: swtaarrs

Differential Revision: D5218067

fbshipit-source-id: 3e97b5876be6b7011377278b77373e7502c7851d