Add Hh_json.json_truncate_string
Summary:
Say you want to dump some json to your logs. That's fine.

But what if it's too long? What if you only want to dump the first 500 bytes or so? And more to the point, *WHAT IF* your logs understand json, and they require well-formed json?

This diff adds a "json_truncate" method to do that.

It's configurable in a bunch of ways -- limit the depth of the json, or limit how many child elements each thing has, or limit the total number of elements emitted, or limit the length of string values.

Reviewed By: arxanas

Differential Revision: D5212977

fbshipit-source-id: 0ea804aca6f8aa63687d8cfd32f2fe1de1bed7d1