log the message-params
Summary:
Parity. The old persistent connection, whenever it shelled out to hh_client, would log argv. That let you observe which files, lines &c. it was working with.

The LSP equivalent is to log the "params" we receive on each request/notification. That's what this diff does.

I wondered whether to log the entirety of the jsonrpc message. That would give us richer logs\\! But in cases like didOpen, it would include the entire text of documents, which I think would be too big. So I'm truncating the json we store into something shorter.

Reviewed By: arxanas

Differential Revision: D5214200

fbshipit-source-id: f316caa2cb91f2a1bc21a47cf8679c04b2f4ba4d