Another tweak to phi opts
Summary:
After my recent diffs for statics, we still ended up with a phi
between two identical LdStaticLocs. I had planned to add an
optimization to gvn to deal with this generically, but just noticed
that optimizePhis already does this for 3 other instructions; this
just adds LdStaticLoc to the set.

Reviewed By: ricklavoie

Differential Revision: D5238212

fbshipit-source-id: 5e71512b625c878def1ec35d75d6e9b276ca80f4