the "hack is busy/ready" dialog auto-dismisses
Summary:
We display the "hack is busy" dialog when Hack takes a long time to start up, and then display "hack is ready" once it's done. With this diff,

(1) the "busy" dialog will automatically be dismissed when the "ready" dialog appears
(2) the "ready" dialog will automatically be dismissed when the user does any actions.

Reviewed By: arxanas

Differential Revision: D5224062

fbshipit-source-id: 95f08865debfa135c5db4eb4dda45aa1b89526f3