*trivial* remove some loggin
Summary:
I went a bit overboard on the logging...

Nuclide is already logging every message it sends+receives.

In addition to that, I'd had "hh_client lsp" write to Nuclide's log for every event it receives (so as to indicate what state hh_client is in, so as to have an entry inside Nuclide's log for events that came from the server, not just from the client). Of course it did that by sending a "window/telemetry" message to the client, so it got logged twice.

But this has been adding too much noise, so I want to remove it.

Reviewed By: arxanas

Differential Revision: D5237704

fbshipit-source-id: e4404fec2753fc633506241c90c437782610ed79