refresh regular labels in fault blocks
Summary: in D5231323 we've started to create fresh copies of finally blocks to avoid having duplicate labels - this had an unfortunate effect of refreshing all labels including faults which was not properly detected if they were nested inside other fault blocks. This diff reverts D5231323 and takes another approach - instead of emitting finally body twice we do it once for normal case and for fault case we create a copy refreshing regular labels that were introduced in this block to make sure they won't collide with normal case.

Reviewed By: oulgen

Differential Revision: D5237262

fbshipit-source-id: 7b86f273d12cea3bb083cd6cc5a51e314691d73a