Fix unnecessary split in FunctionStaticStatement with one initializer
Summary:
This diff fixes the formatting of FunctionStaticStatements with only one initializer, where the initializer contains a line break. Previously, they would be formatted like this:

```
function foo() {
  static
    $my_dict = dict[
      'foo' => 'long foo value to take up space',
      'bar' => 'long bar value to take up space',
    ];
}
```

This is because the `dict` literal needs to be split up onto multiple lines, which causes the split before `$my_dict` to break, because that split's rule was configured to care about its child rules. This is the right behavior when we have more than one initializer:

```
function foo() {
  static
    $my_dict = dict[
      'foo' => 'long foo value to take up space',
      'bar' => 'long bar value to take up space',
    ],
    $my_string = 'baz';
}
```

But it looks broken when there's only one initializer. The split before a lone initializer shouldn't depend on whether its child rules break.

We actually solved this problem already for class-level static initializers, so this diff just moves that transformation into a function (`handle_declarator_list`) and reuses it for "keyword_expr_list_statements".

Reviewed By: arxanas

Differential Revision: D5241482

fbshipit-source-id: db14fb72e44cf7a05a508e5a5221da1bf3944115