Return only a SyntaxTree from parse fn
Summary: Previously, `parse` returned a tuple of `(SourceText, SyntaxTree, EditableSyntax)`, since these things are generally all needed by hackfmt's various formatting functions. However, a `SyntaxTree` contains a reference to its `SourceText`, and it's a one-liner to produce an `EditableSyntax` from a `SyntaxTree`, so we can simplify `parse` (and the functions downstream of it) by just returning a `SyntaxTree`.

Reviewed By: arxanas

Differential Revision: D5243358

fbshipit-source-id: 6b5b4da5016e2ef23baff4b3de51c12856e0c1a6