Don't stop diff formatting because of parse errors
Summary:
When formatting the changed lines of a diff, hackfmt is intentionally conservative, and tries to bail *before* writing to any files if it detects that anything is wrong. This is because it only reads the filenames and line intervals from a diff before formatting those lines--if the diff is out of date, we could end up formatting completely incorrect files or ranges. Since we encourage people to use this tool on uncommitted changes in their working tree (`hg diff | hackfmt --diff`), trashing it could mean losing work, so we try to tread carefully.

This behavior becomes a problem when formatting a very large diff (e.g. a codemod). If we encounter a parse error in some file, that might be a signal that there's something wrong with the codemod (so we should communicate that there was a problem), but there's no need to bail on formatting the rest of the files.

This change allows --diff mode to proceed when one of the files to be formatted has a parse error. When we encounter a file with a parse error, we print a warning to stderr.

Depends on D5243358

Reviewed By: arxanas

Differential Revision: D5243372

fbshipit-source-id: 8c4e71a343b9495f9fea6f6c0170c185df9914a6