WIP: Keyword FFP-based Autocomplete
Summary:
Now that we've finished the testing framework and can extract the
location that we wish to autocomplete at, we can attempt to auto-complete some
keywords.
The general strategy for doing so is as follows:
1. We extract the location from a comment at the top of the test file as before
2. We retrieve the leaf node at this location
3. We get the full tree traversal path from root to the leaf node
4. We analyze this path in order to determine certain information about the context that our leaf node is in, i.e. are we inside the toplevel or inside a function
5. We return a list of keywords that are valid given the current context

Currently, keywords that can be determined purely based on the path from the
root of the tree to the leaf node can generally be autocompleted. Keywords that
require information about their siblings in the syntax tree are still WIP.

For an example of a completion that requires sibling information, consider the following snippet:

```
<?hh

function foo(){
  if (condition){
    doThings();
  }
  el
```

Clearly, we wish for the "el" to be completed with either "elseif" or "else".
Unfortunately, the parse tree sees "el" as the beginning of a new statement and
not as part of the if statement. Therefore, more information than just the
direct path to the node is required here to do this completion.

Reviewed By: arxanas

Differential Revision: D5221442

fbshipit-source-id: b74ab9ae99e627c941fd5b9041e143c457951812