Catching 'implements' keyword error
Summary: I added a case to the ClassishDeclaration arm of the match statement in classish_errors in full_fidelity_errors, in order to catch the error resulting from improper use of the 'implements' keyword in a classish declaration. (A user may only use 'implements' in a class declaration, but previously the parser was allowing it in trait and interface declarations as well.) I also added SyntaxError.error2035 to address this issue. (In extending classish_errors, I tried to mimic the structure from the MethodishDeclaration branch of methodish_errors.)

Reviewed By: ericlippert

Differential Revision: D5222087

fbshipit-source-id: 0b01a95a826b5b2061db6447743f5973a7ad84b7