Add --inheritance-ancestor-classes, --inheritance-ancestor-interfaces, and --inheritance-ancestor-traits options
Summary:
Add these options for HackAST.

I considered using the same `--inheritance-ancestors` flag and just passing an additional flag determining the type of thing to filter, but there didn't seem to be any precedent for some of the flags taking additional, general-purpose arguments. This also is a solution that maintains backwards-compatibility.

Differential Revision: D5250324

fbshipit-source-id: 43487c2d9711f34b8df432e40249df315b341d2d