Make split after lambda arrow independent
Summary:
Previously, the split after a lambda arrow was dependent on the splits inside the lambda body, which meant that it would always be broken on if there were any breaks in the lambda body (which is probably the case more often than not).

This change just makes that split independent, so that the body can begin on the same line as the lambda arrow even if the body is multi-line (like a function call with line breaks between the arguments).

Reviewed By: arxanas

Differential Revision: D5253660

fbshipit-source-id: e66706bde28f9347bd0e72cfe286853335e98b30