switch LSP to file urls
Summary:
LSP says that the URIs it receives and sends are expected to be file URLs. This diff does that.

Actually, it's a bit flexible on the receiving side, allowing both filenames and file URLs to be passed on. I have a hunch that some clients might work this way. But on the sending side, it always emits file URLs.

Reviewed By: arxanas

Differential Revision: D5240108

fbshipit-source-id: ddcb993d1b2ee392d128fdd176e76d8896a4d611