*easy* Return 'filename:string' rather than 'ServerUtils.FileName'
Summary:
This diff is in preparation for further work. There's a helper function which used to return 'ServerUtils.FileName filename'. Now it just returns a string 'filename'.

The ultimate reason is that hh_server sometimes returns empty filepaths. We're going to need the string filename so we can use it as a default in that case.

Reviewed By: arxanas

Differential Revision: D5240142

fbshipit-source-id: d6c66ae4b6eca25dc5d97cbb393cf40ea9e9d5e7