align lsp naming convention with LSP spec
Summary:
The Ministry of War sent out an urgent call: war effort needs more underscores\\! We at the home front have to do all we can to support the effort by giving up our underscores. Those two underscores in `text_document_identifier`, they'll be used to make ammunition for the lads. The underscore in `Message_type` will be used to make a plane strut.

This diff isn't settled -- I'm sending it out to get your opinions. What do you think?

It's something that's been bugging me for a while. I want to use the same naming conventions as the LSP spec as much as possible. This will mean that (1) when we grep our codebase for something related to LSP, we'll find the relevant OCaml parts as well as the relevant Nuclide parts; (2) there'll be fewer bugs like one Waleed found where we'd serialized something into JSON using the underscore name rather than the proper JSON name; (3) it will reduce the mental overhead of switching between the LSP spec and the OCaml implementation.

Reviewed By: arxanas

Differential Revision: D5244803

fbshipit-source-id: 62f5de9debad1d558e9f64313105a4c463fe5eee