immediate mutation should now preserve stack correctness
Summary: The mutate_immediates function should now never create a situation in which the file fails assembly because of a stack depth mismatch. When an operation requires more cells than are currently on the stack, cells trivially filled with values of the proper flavor are added to the stack, and if the stack is too tall before a control flow instruction, Pop* instructions of the proper flavor are inserted as necessary.

Reviewed By: paulbiss

Differential Revision: D5231305

fbshipit-source-id: dd453975f1d10b75b9132695a8549b25a6731459