Hack codegen: semantic diff in test framework (not enabled by default)
Summary:
This is a first cut at integrating the semantic diff tool into our testing framework for Hack code generation.
The tool is run against the output of HHVM and `hh_single_compile`. If it fails (e.g. raises an exception or
times out) then we drop through to vanilla diff. If it succeeds, we look at the output of the tool, and if the distance is
zero (i.e. the outputs of HHVM and Hack are deemed equivalent) then we regard the test as passing and `diff` is not applied.
If the distance is non-zero, we run vanilla diff.

The test runner does not run semdiff by default: you must specify the `--semdiff` option. The TARGETS file includes this option for hhcodegen tests.
Many new tests in slow and quick now pass - verify by manual diff that semdiff is doing its job and that the differences are not observable.

Reviewed By: paulbiss

Differential Revision: D5137818

fbshipit-source-id: dfa13f1184b66fb44431f143121dccfa2d230995