Duplicate and remove mutations now also preserve stack
Summary: The duplicate, remove and immediate mutation types should now honor the stack, in the sense that the stack height at the beginning and end of basic blocks will be the same as they were in it was in the original program. This makes no promises about the types, flavors or values present in the stack.

Reviewed By: paulbiss

Differential Revision: D5260802

fbshipit-source-id: c45562ac02b4c8b21010154b02219487dca62d12