Add a weight field to Vasm blocks
Summary:
The code layout pass was computing weights for Vasm blocks, and this diff moves
the weight into a Vblock field.  This will allow us to set the block weights
differently based on profile information collected for Vasm-level blocks.  I
plan to do that separately though, and this diff shouldn't affect any behavior.

Reviewed By: paulbiss

Differential Revision: D5269471

fbshipit-source-id: 20da8388862920345c4bf3344262d029aa268ed7