Pass return state via return registers
Summary:
When generators are iterated, ContEnter is responsible for putting value
"returned" by generator's Yield, Await or RetC opcode on the stack. This works
nicely in the interpreter, since the returned value is already set to the right
stack position by the generator.

On the Yield/RetC side, JIT used both return registers and return via stack. On
the ContEnter side, there was simply no logic and we were relying on JIT to
load the value from the stack. This didn't work properly in some corner cases
involving a combination of JIT and interpreter and we ended up returning null,
which was on the same stack position prior to ContEnter.

This didn't cause any issues, since Generators always return null anyway and
AsyncGenerators did not have proper JIT support, never exercising this code
path.

Let's move propagation of return values to return registers and push the
returned value to the stack explicitly from ContEnter.

Reviewed By: paulbiss

Differential Revision: D5273252

fbshipit-source-id: e8c50739fe871e74c87f724dd3a035d4ef94ad4a