Fix breaking in catch clause
Summary:
Previously, CatchClause was using Splits, but without a containing `WithRule` block, so its splits would end up depending on the nearest containing rule. This meant that if the containing rule was broken (as it is in the added test case `try_catch_in_broken_parent_rule`), the splits around the CatchClause's exception variable would break too.

This change adds the missing WithRule block and allows splitting between the exception type and variable.

Reviewed By: arxanas

Differential Revision: D5267011

fbshipit-source-id: 2b0b1ae1ca4ac47fb86a434a5020da5d1143c08a