Run parser on stdin, dump result
Summary:
Implements a "compiler" that can dump the AST we get from running the Zend
parser; mostly to verify that this is working.

I'm not sure if it's worth it to unit test the exact parser output since this
is likely to change at least slightly as the upstream parser does.

Reviewed By: mofarrell

Differential Revision: D5270508

fbshipit-source-id: 541ffa3a45bcc1ba281c9141f1525ecb4082fd2d