Fix a few issues with hfsort collection process
* Pass --ffunctions-sections to C code in third-party modules. Otherwise
the linker will not find hot C functions if they are referenced in the
linker script.
* Only collect user space stacks in hf-prod-collect. Since hfsort only
looks at the topmost two functions on the stack, having kernel functions
there will cause less accurate information to be used.
* Add possibility to pass HHVM_PID to hf-prod-collect through an env
var. This is to account for the case where we use oss-performance to
train. When we do that, 'pgrep -o hhvm' matches the HHVM process that
runs perf.php instead of the HHVM under test.
* Regenerate linker script to include all these changes. oss-performance
WordPress was uses as the training workload.