Implement Hack equality function naively
Summary:
The goal of this diff is to implement an equality function that checks if two types are exactly equal, with something cheaper than unification. This function does it naively, which may be too slow for our needs, but I wanted to get something out to show the idea.

We want to implement a function ty_equal such that if ty_equal ty1 ty2, then ty1 and ty2 unify. The converse is not necessarily true.

Differential Revision: D4998171

fbshipit-source-id: f40af46caf4f40fd69c599f2e6b8fbcef7e2e57a