HHBBC shouldn't optimize away CastVArray on arrays
Summary:
The CastVArray handling in HHBBC used the same casting helper as the rest, which
meant that HHBBC could optimize away a CastVArray operation if the input was
already an array. This isn't correct since CastVArray isn't a no-op on arrays
(it drops the keys).

While I'm here, add some Hack array handling that was missing from CastBool and
AKExists. The later was not accounting for HackArrCompatNotices possibly raising
warnings.

Reviewed By: mxw

Differential Revision: D5281705

fbshipit-source-id: f018ca20307a0ae8287d592a238d414658c17c41