Adding support for keywords async and await
Summary:
Changes:
- Added support for `async` and `await` with tests for this functionality
- Deleted old `get_child_and_predecessor_at_pos` function
- Autocomplete results are now sorted in alphabetical order

Since the keywords `async` and `await` are only valid inside async functions, we need a way to determine if we are inside an asynchronous function. However, this information is within the function header in the parse tree, which is not directly on the path from root to the node inside the function body.

Therefore, I modified the code such that the syntax tree nodes on the path from root to node are passed directly to the keyword complete function instead of only passing the kinds of the nodes as we were doing previously.

I also added a field to the context object to track this information.

I also realized that if the parse tree sees a spot where the keyword "insteadof" is supposed to go in a use expression, the parse will throw an error instead of putting a missing node in that location.

Reviewed By: arxanas

Differential Revision: D5259832

fbshipit-source-id: 66e2c0999a1c6ce8bd3d4bd493a2f077282cc500