Add an option to hh single compile to read config from CLI
Summary:
Add an option to hh single compile to read config from CLI

Now that the hack codegen project is becoming more mature, we need to be able to share the config file with HHVM. In order to do so, we are adding two options to hh single compile. First one is to read a config file passed on the command line by `-c config.json`. The second one, not implemented on this diff but will be done by paulbiss, is to read a config file that is passed on stdin when HackC is operating in DAEMON mode.

Reviewed By: paulbiss

Differential Revision: D5278134

fbshipit-source-id: 36915ae7b2bfb93020ff89118d4aa73b5c307608