Add autoload_map target to DynamicFactsDB generation step so that Hack tracks it
Summary:
Using DynamicFactsDB here doesn't include build target of flib/autoload/autoload_map.php, so we don't end up parsing it when we initialize the server. This causes a crash if it gets deleted under our feet.

I'm not sure this is actually the best approach to doing this change, but it should definitely solve the crashing problem.

Reviewed By: periodic1236

Differential Revision: D5289655

fbshipit-source-id: 1c25a479acc8bdecf3173561a2e54687dd502a9a