Start emitter in earnest and bytecode bookkeeping
Summary:
Start laying groundwork for the actual compilation machinery. Importantly, I
added a separate set of types to store bytecode and metadata at the level of
fidelity I need (not enough to roundtrip, just enough to emit correct hhas). I
replicated the approach in `hhbbc/representation.h` and `hhbbc/bc.h` where each
opcode is stored with its immediates in its own structure type and bytecode in
general is a tagged union of those structs.

Obviously not all the metadata is being stored yet (nor do I handle anything
but straightline functions so far) so this is certainly very likely to change

We can compile some simple expressions, though. Number literals, addition, and
echo work; though the bytecode is not 100% correct since I'm not yet tracking
stack depth.

Next steps are to allow HHVM to call into our binary so I can add proper tests
to HHVM's suite and to flesh out the design of the compiler itself and bytecode
representations.

One possible issue with the current design is stack usage in the compiler. I
plan to reorganize things so it will run in constant stack space (probably with
liberal use of closures)

Reviewed By: mofarrell

Differential Revision: D5296023

fbshipit-source-id: 294a04088dfa4eba01dc1c0e0a70c178c7523d22