Provide mechanism in repo to toggle into faked Watchman Event Watcher behavior
Summary:
This creates a mechanism for the repo itself to toggle the Watchman Event Watcher
behavior (disable it and have it always spit out a fake "Settled" message).

A bug was actually fixed in D5185200 which would result in log-spam until hard
disk was full if a Hack client disconnected while it was in purgatory and the
Monitor attempted to send it a message later (the summary for that diff doesn't
mention fixing this bug - it fixed another bug - but it works). Purgatory usage was rare
but becomes common with use of the Informant with a Watchman Event Watcher
running.

To avoid this, we need to disable the Watchman Event Watcher for older versions
of the Hack server. Of course this isn't really possible by making changes to
Hack itself. Here I add a file which acts as a toggle to enable/disable the
Watchman Event Watcher. When disabled, all messages are faked into a Settled
message instead.

Reviewed By: dlreeves

Differential Revision: D5298716

fbshipit-source-id: 25cce95a3dcb9a2f97260c800b985667c9150ee7