added checking for a couple fuzzer bugs to assembler/verifier
Summary:
The fuzzer turned up two small errors in the verifier, allowing crashing programs through.

First, we were not explicitly checking that IVA immediates are non-negative. This actually crashed the verifier itself, so this is now checked in the assembler.

Second, we weren't checking that the stack offset in member keys in final member ops was actually a valid stack cell, so a check for this was added to the verifier.

Reviewed By: paulbiss

Differential Revision: D5288173

fbshipit-source-id: c973d51d4c6b3237abf142936e879251c64a0853