Add command-line flag to explicitly provide filename for logging
Summary:
This flag lets you specify the filename for logging purposes. In particular, this is useful for commands that take their input on stdin, because we don't have the filename in that case. (Those files typically contained the modified, unsaved contents inside the IDE, so they don't have a real backing file on disk.)

This diff involves a refactoring to make the various commands use a record type instead of a tuple. This is to prevent confusion not catchable by the type system between the filename and the filename for logging. This record type is implemented as a module rather than just a flat record because otherwise the compiler warns about type-directed disambiguation for the record type, because there are multiple record types that use identically-named fields.

Note that as written, the `filename` and `filename_for_logging` variables can't be collapsed into one variable. This is because we use the absence of a `filename` to mean that we should read from stdin, but we will probably provide a filename for logging in that case.

Differential Revision: D5193807

fbshipit-source-id: 235de06d5eec030bf4d0480c60feaed9c24bd22c