*easy* provide kind+details for autocomplete
Summary:
The previous implementation of autocomplete was a bit rough and ready. This diff has two changes:

(1) Kind. The previous persistent connection code only knew about the 'function' kind. This diff brings us up to parity. For future work, we'll get more kinds.

(2) Detail. The previous persistent connection code just returned the type of the thing (be it method or variable or class). We bring it up to parity. For future work, we'll return the proper signature.

This diff also adds a bunch of TODO comments for future work.

Reviewed By: arxanas

Differential Revision: D5302774

fbshipit-source-id: 4e8045e0b2a9dc056ee70ae666fea74a8dc44a79