Fixing bug with parameter extraction in lambda_analyzer.ml
Summary:
The function syntax_node_to_list in full_fidelity_syntax.ml extracts
the underlying list of items from a SyntaxList. This list turns out to be a
list of "ListItems", which then contain the actual nodes within the list. This
fix properly accounts for the ListItem wrappers when extracting parameters from
functions.

Reviewed By: ericlippert

Differential Revision: D5294340

fbshipit-source-id: e3efa5f84e99eb232ab689bc774c1e1c44d03937