Call move_strong_iterators in the right places.
Summary:
Array copies have distinct semantics when the array in question is being iterated using foreach by reference; move_strong_iterators can be used to hide array copies that should not trigger foreach by reference's array copy semantics.
To match PHP's behavior, we need to do this when growing packed or mixed arrays (otherwise the grow operations look like copies), but not if we've copied them first.
Remove misleading comments about where this is needed, move the calls, and add a test for this behavior.
Also fix the signature of move_strong_iterators and make it inlineable.

Reviewed By: jano

Differential Revision: D5304268

fbshipit-source-id: abebc79af3613aace7980c6bcb8afe1d6e65c4a7