Emit function body dispatch code in the proper code area
Summary:
The JIT was always emitting the function body dispatch code in the main code
area, because no translation kind was passed to CodeCache::view() when emitting
them.  This diff fixes the issue by properly passing the corresponding
translation kind around, allowing the body dispatches to be placed in either
main, prof, or hot code areas.  Also, for when a function gets re-JITed in
optimized mode, we now re-emit the function body dispatch immediately before
that function's optimized prologues.

Reviewed By: paulbiss

Differential Revision: D5313015

fbshipit-source-id: bbc1d1249acab4e5b081048e848a25421e8e2886