Ensure ExpressionList state is set correctly for varrays
Summary:
Varray literal constructors use the same grammar as vec, but produce PHP
arrays. The vec grammar produces lists of Expressions, while the PHP array
grammar produces lists of ArrayPairExpressions. Since the varray literal
produces PHP arrays, the parser turns the list of Expressions into a list of
ArrayPairExpressions (so everything else just works). However this logic
attempted to modify the ExpressionList node in place, which is wrong, because
other state in the ExpressionList has already been set. Instead of changing the
ExpressionList in place, create a new one, which ensures the other state will be
set correctly.

Reviewed By: paulbiss

Differential Revision: D5313556

fbshipit-source-id: 78c1fb1a828bd4a711e265e7cc671fd6d3126d17