Compact backtrace should check for empty array before calling MakeUncounted
Summary:
PackedArray::MakeUncounted (and family) will assert if given an empty
array. So, the compact backtrace generation code should refrain from using it if
the array it has is empty (and just use the static empty array instead). This
could result in double asserts, where we'll assert on something, try to generate
a backtrace from the assert handler, and then assert again.

Reviewed By: alexeyt

Differential Revision: D5313891

fbshipit-source-id: a87439f8bc85da12964b0895a9609fddad14e606