Add missing features to hhas/hackc and enable new round-trip tests
Summary:
This diff adds a variety of missing features to hhas as either new directives or extensions to existing directives, these include

 - Proper support for statics (via `.static`)
 - Type structures attached to aliases (via a new optional component of the `.alias` directive)
 - Support for annotating which functions and classes are not top level and therefore not possibly hoistable via a new `nontop` attribute
 - Gated support for doc-comments (enabled via `Eval.DisassemblerDocComments`, and defaulted on)
 - Gated support for line number/source position mapping (via `.srcloc` directive and optional line pairs on class and function declarations, enabled via `Eval.DisassemblerSourceMapping`)
 - Support for loading filepaths from the bytecode repo (required to get round tripped testes to pass, enabled via `Eval.LoadFilepathFromUnitCache` and only used for testing)

Additionally changes were made to the way that the assembler handles unreachable code to be more inline with the hphpc emitter. In particular all instructions will now unconditionally force themselves to be reachable with an initial stack depth of 0. This isn't an ideal solution but it most closely matches what we do elsewhere which should ease writing new code generators as the existing frontend functioned this way. It was also necessary to force the hphpc frontend to assume that the bytecodes following an Exit instruction were unconditionally reachable as it was only sometimes eliding pops confusing the assembler and breaking round-trip tests.

These new features have all been implemented in hackc as well, with the exception of doc comments and source mapping, which will come in follow-up diffs. Notably, these features exposed a few latent bugs that had been missed by our incomplete disassembler, these tests are now marked as failing (a handful of tests are newly passing). Many of the ~30 tests marked as failing in this diff are semantically equivalent, however, the phpCode encoded alongside `.static` directives was not always identical.

Reviewed By: andrewjkennedy

Differential Revision: D5291798

fbshipit-source-id: 20f20d97c00502003536862e423e1af85564dad7