Object iteration does not need to be handled in every code path.
Summary:
The strong iterator code was written too generically, which made it hard to reason about.
The way object iteration is currently implemented, we always have a mixed array with refcount 1 that no other code can ever get a reference to or replace with something else.
This means we never need to escalateCheck, cowCheck, or reregister if we're strong iterating an object.
It also means advance and prepare can do much less work in the hasAd (object iteration) case.
Finally, constructors (for both object and array iteration) don't need to handle data == nullptr; the callers always ensure that we have something valid to iterate.

Reviewed By: jano

Differential Revision: D5312147

fbshipit-source-id: a35190b6d5a9731cd4118899a4a05bda2d2eae21