Allow order of free variables to be permuted when comparing closure classes
Summary:
The order in which free variables are passed in as properties when a closure is
created is not significant. So we match a pattern of CUGetL's followed by CreateCl on
both sides and if what's passed is a permutation of named variables then we record the
permutation and schedule the classes for comparison modulo that permutation being
applied to their properties.

Reviewed By: andrewjkennedy

Differential Revision: D5300796

fbshipit-source-id: 0232aa58a5e2fea300aea9eb53c18826570bd86b